+++ src/index.ts
@@ -1,5 +1,4 @@
-import FlipCardPanel from './FlipCardPanel';
 export * from './types';
-export { default as FlipCard } from './FlipCard';
-export { default as OdometerCard } from './OdometerCard';
-export default FlipCardPanel;
+export { FlipCard } from './FlipCard';
+export { OdometerCard } from './OdometerCard';
+export { FlipCardPanel } from './FlipCardPanel';